requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1; python_version < "3.9"
Django==4.2.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
psycopg2-binary==2.9.9
